feat: add isolated (non-singleton) OpenFeature API instances
- Add createIsolatedOpenFeatureAPI() via /isolated sub-path export for web and server SDKs
- Extract core logic into OpenFeatureAPIBase class, keeping OpenFeatureAPI as thin singleton wrapper
- Update React SDK OpenFeatureProvider to accept optional openfeature prop for isolated instances
- Add comprehensive tests for isolation of providers, context, hooks, and event handlers

Addresses singleton limitations for micro-frontend architectures and testing scenarios.
See: https://github.com/open-feature/spec/issues/359
