+++ packages/react/src/provider/provider.tsx
@@ -1,5 +1,5 @@
 import { withFrameworkMetadata } from '@openfeature/core';
-import type { Client } from '@openfeature/web-sdk';
+import type { Client, OpenFeatureAPIBase } from '@openfeature/web-sdk';
 import { OpenFeature } from '@openfeature/web-sdk';
 import * as React from 'react';
 import type { ReactFlagEvaluationOptions } from '../options';
@@ -12,6 +12,28 @@ type ClientOrDomain =
        * @see OpenFeature.setProvider() and overloads.
        */
       domain?: string;
+      /**
+       * An isolated OpenFeature API instance to use instead of the global singleton.
+       * Use this in micro-frontend architectures where different parts of the application
+       * need isolated OpenFeature instances.
+       * @see createIsolatedOpenFeatureAPI from '@openfeature/web-sdk/isolated'
+       * @example
+       * ```tsx
+       * import { createIsolatedOpenFeatureAPI } from '@openfeature/web-sdk/isolated';
+       *
+       * const MyOpenFeature = createIsolatedOpenFeatureAPI();
+       * MyOpenFeature.setProvider(myProvider);
+       *
+       * function App() {
+       *   return (
+       *     <OpenFeatureProvider openfeature={MyOpenFeature}>
+       *       {children}
+       *     </OpenFeatureProvider>
+       *   );
+       * }
+       * ```
+       */
+      openfeature?: OpenFeatureAPIBase;
       client?: never;
     }
   | {
@@ -20,6 +42,7 @@ type ClientOrDomain =
        */
       client?: Client;
       domain?: never;
+      openfeature?: never;
     };
 
 type ProviderProps = {
@@ -32,10 +55,10 @@ type ProviderProps = {
  * @param {ProviderProps} properties props for the context provider
  * @returns {OpenFeatureProvider} context provider
  */
-export function OpenFeatureProvider({ client, domain, children, ...options }: ProviderProps) {
+export function OpenFeatureProvider({ client, domain, openfeature, children, ...options }: ProviderProps) {
   const stableClient = React.useMemo(
-    () => withFrameworkMetadata(client || OpenFeature.getClient(domain), 'react'),
-    [client, domain],
+    () => withFrameworkMetadata(client || (openfeature ?? OpenFeature).getClient(domain), 'react'),
+    [client, domain, openfeature],
   );
 
   return <Context.Provider value={{ client: stableClient, options }}>{children}</Context.Provider>;
