+++ packages/server/src/index.ts
@@ -2,6 +2,7 @@ export * from './client';
 export * from './provider';
 export * from './evaluation';
 export * from './open-feature';
+export * from './open-feature-base';
 export * from './transaction-context';
 export * from './events';
 export * from './hooks';
