+++ packages/web/package.json
@@ -9,10 +9,18 @@
     "dist/"
   ],
   "exports": {
-    "types": "./dist/types.d.ts",
-    "import": "./dist/esm/index.js",
-    "require": "./dist/cjs/index.js",
-    "default": "./dist/cjs/index.js"
+    ".": {
+      "types": "./dist/types.d.ts",
+      "import": "./dist/esm/index.js",
+      "require": "./dist/cjs/index.js",
+      "default": "./dist/cjs/index.js"
+    },
+    "./isolated": {
+      "types": "./dist/isolated.d.ts",
+      "import": "./dist/esm/isolated.js",
+      "require": "./dist/cjs/isolated.js",
+      "default": "./dist/cjs/isolated.js"
+    }
   },
   "types": "./dist/types.d.ts",
   "scripts": {
@@ -21,11 +29,14 @@
     "lint:fix": "eslint ./ --fix",
     "clean": "shx rm -rf ./dist",
     "build:web-esm": "esbuild src/index.ts --bundle --external:@openfeature/core --sourcemap --target=es2015 --platform=browser --format=esm --outfile=./dist/esm/index.js --analyze",
+    "build:web-esm-isolated": "esbuild src/isolated.ts --bundle --external:@openfeature/core --sourcemap --target=es2015 --platform=browser --format=esm --outfile=./dist/esm/isolated.js --analyze",
     "build:web-cjs": "esbuild src/index.ts --bundle --external:@openfeature/core --sourcemap --target=es2015 --platform=browser --format=cjs --outfile=./dist/cjs/index.js --analyze",
+    "build:web-cjs-isolated": "esbuild src/isolated.ts --bundle --external:@openfeature/core --sourcemap --target=es2015 --platform=browser --format=cjs --outfile=./dist/cjs/isolated.js --analyze",
     "build:web-global": "esbuild src/index.ts --bundle --sourcemap --target=es2015 --platform=browser --format=iife --outfile=./dist/global/index.js --global-name=OpenFeature --analyze",
     "build:web-global:min": "esbuild src/index.ts --bundle --sourcemap --target=es2015 --platform=browser --format=iife --outfile=./dist/global/index.min.js --global-name=OpenFeature --minify --analyze",
     "build:rollup-types": "rollup -c ../../rollup.config.mjs",
-    "build": "npm run clean && npm run build:web-esm && npm run build:web-cjs && npm run build:web-global && npm run build:web-global:min && npm run build:rollup-types",
+    "build:rollup-types-isolated": "rollup -c ./rollup.isolated.config.mjs",
+    "build": "npm run clean && npm run build:web-esm && npm run build:web-esm-isolated && npm run build:web-cjs && npm run build:web-cjs-isolated && npm run build:web-global && npm run build:web-global:min && npm run build:rollup-types && npm run build:rollup-types-isolated",
     "postbuild": "shx cp ./../../package.esm.json ./dist/esm/package.json",
     "current-version": "echo $npm_package_version",
     "prepack": "shx cp ./../../LICENSE ./LICENSE",
